requirements.txt
anyio==3.7.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.96.0
h11==0.14.0
idna==3.4
pydantic==1.10.8
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
SQLAlchemy==2.0.16
psycopg2==2.9.5
strawberry-graphql==0.186.2